Use gitattributes to manage generated files
I want the viewable images to be available, but, I don't want to see the
diffs in the SVG.  That is ultimately not useful.  Instead, I can [set
the attributes][ref] to make the SVG appear to be a binary file.  Now we
will just take the newest version as the "right" one.  (In fact, I
should do this for my website too).

[ref]: https://medium.com/@clarkbw/managing-generated-files-in-github-1f1989c09dfd>